refactor(tokenizer): keep json.dumps helper minimal and rename to json_dumps
serde_json ships only compact and pretty formatters, neither matching
Python's single-line `json.dumps` spacing, so a small custom formatter is
needed. Keep it as a focused 3-override formatter in a dedicated
`json_dumps` module (`json_dumps::to_string`) rather than relocating
chat_template's configurable HF-tojson formatter — the two have different
needs (fixed default vs configurable), and a move would be pure churn.
chat_template is left untouched.
