+++ crates/tokenizer/src/lib.rs
@@ -7,8 +7,8 @@ pub mod encoders;
 pub mod eos;
 pub mod factory;
 pub mod hub;
+pub(crate) mod json_dumps;
 pub mod mock;
-pub(crate) mod python_json;
 pub mod registry;
 pub mod sequence;
 pub mod stop;
